DescribedEntity moved to 'client.model.core'
