+++ src/main/java/egl/client/model/core/DescribedEntity.java
@@ -1,9 +1,8 @@
-package egl.core.model;
+package egl.client.model.core;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 
-import egl.client.model.core.DatabaseEntity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
