+++ src/main/java/egl/core/model/topic/Topic.java
@@ -3,7 +3,7 @@ package egl.core.model.topic;
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
 
-import egl.core.model.DescribedEntity;
+import egl.client.model.core.DescribedEntity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
